feat(studio): move landing page to root
